Add code to run locally Fixes #7
To be able to run locally this project was dependant on
GOB-Message-Broker, whose only purpose was the provide functionality to
run GOB-Workflow (and other components) locally.

This commit adds code from that project to this one, so that it can be
run locally without that project. The GOB-Message-Broker can be
discontinued.