Auto reconnect wrapper
A decorator that automatically reconnects when execution
fails due to connection problems.

The decorator can handle any connections, as long as a
connect, disconnect and is_connected function are provided

The decorator should be moved to GOB-Core when accepted
and then also be used for the communication with the
message broker.

The current implementation is for GOB Workflow only and
serves a starting point for moving it to GOB Core

Coverage set to 100%